refactor(api): migrate group route  to its bounded context
