+++ api/lib/application/campaigns/campaign-controller.js
@@ -1,6 +1,5 @@
 import * as campaignAnalysisSerializer from '../../../src/prescription/campaign-participation/infrastructure/serializers/jsonapi/campaign-analysis-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
-import * as groupSerializer from '../../infrastructure/serializers/jsonapi/group-serializer.js';
 import { extractLocaleFromRequest } from '../../infrastructure/utils/request-response-utils.js';
 
 const getAnalysis = async function (request, h, dependencies = { campaignAnalysisSerializer }) {
@@ -12,17 +11,8 @@ const getAnalysis = async function (request, h, dependencies = { campaignAnalysi
   return dependencies.campaignAnalysisSerializer.serialize(campaignAnalysis);
 };
 
-const getGroups = async function (request) {
-  const { userId } = request.auth.credentials;
-  const campaignId = request.params.id;
-
-  const groups = await usecases.getParticipantsGroup({ userId, campaignId });
-  return groupSerializer.serialize(groups);
-};
-
 const campaignController = {
   getAnalysis,
-  getGroups,
 };
 
 export { campaignController };
