+++ api/lib/application/campaigns/index.js
@@ -25,23 +25,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/campaigns/{id}/groups',
-      config: {
-        validate: {
-          params: Joi.object({
-            id: identifiersType.campaignId,
-          }),
-        },
-        handler: campaignController.getGroups,
-        notes: [
-          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
-            '- Récupération des groupes des participants à la campagne',
-        ],
-        tags: ['api', 'group'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/campaigns/{id}/stats/participations-by-stage',
       config: {
         validate: {
