+++ api/src/prescription/campaign/application/campaign-controller.js
@@ -1,5 +1,6 @@
 import { usecases } from '../domain/usecases/index.js';
 import * as divisionSerializer from '../infrastructure/serializers/jsonapi/division-serializer.js';
+import * as groupSerializer from '../infrastructure/serializers/jsonapi/group-serializer.js';
 
 const division = async function (request) {
   const { userId } = request.auth.credentials;
@@ -9,8 +10,17 @@ const division = async function (request) {
   return divisionSerializer.serialize(divisions);
 };
 
+const getGroups = async function (request) {
+  const { userId } = request.auth.credentials;
+  const campaignId = request.params.id;
+
+  const groups = await usecases.getParticipantsGroup({ userId, campaignId });
+  return groupSerializer.serialize(groups);
+};
+
 const campaignController = {
   division,
+  getGroups,
 };
 
 export { campaignController };
