+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/group-serializer.js
