+++ api/src/prescription/organization-learner/application/sup-learner-list-controller.js
@@ -1,4 +1,4 @@
-import * as groupSerializer from '../../../../lib/infrastructure/serializers/jsonapi/group-serializer.js';
+import * as groupSerializer from '../../../../src/prescription/campaign/infrastructure/serializers/jsonapi/group-serializer.js';
 import * as queryParamsUtils from '../../../shared/infrastructure/utils/query-params-utils.js';
 import { usecases } from '../domain/usecases/index.js';
 import * as supOrganizationParticipantsSerializer from '../infrastructure/serializers/jsonapi/sup-organization-participants-serializer.js';
