+++ api/tests/unit/application/campaign/index_test.js
@@ -31,33 +31,6 @@ describe('Unit | Application | Router | campaign-router ', function () {
     });
   });
 
-  describe('GET /api/campaigns/{id}/groups', function () {
-    it('should return 200', async function () {
-      // given
-      sinon.stub(campaignController, 'getGroups').callsFake((request, h) => h.response('ok').code(200));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const result = await httpTestServer.request('GET', '/api/campaigns/1/groups');
-
-      // then
-      expect(result.statusCode).to.equal(200);
-    });
-
-    it('should return 400 with an invalid campaign id', async function () {
-      // given
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const result = await httpTestServer.request('GET', '/api/campaigns/invalid/groups');
-
-      // then
-      expect(result.statusCode).to.equal(400);
-    });
-  });
-
   describe('GET /api/campaigns/{id}/stats/participations-by-stage', function () {
     it('should return 200', async function () {
       // given
